:ref: moved qr code builder to common utils, as it can live separately from mfa functionality
