+++ commons/pom.xml
@@ -50,6 +50,11 @@
         <!-- -->
         <!-- External dependencies -->
         <dependency>
+            <!-- QR code generation -->
+            <groupId>com.google.zxing</groupId>
+            <artifactId>javase</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
