+++ commons/src/main/java/org/eclipse/kapua/commons/util/qr/QRCodeBuilder.java
@@ -1,0 +1,19 @@
+/*******************************************************************************
+ * Copyright (c) 2020, 2022 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Eurotech - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.kapua.commons.util.qr;
+
+import java.net.URI;
+
+public interface QRCodeBuilder {
+    String generateQRCode(URI uri);
+}
