+++ commons/src/main/java/org/eclipse/kapua/commons/util/qr/QRCodeBuilderImpl.java
@@ -1,0 +1,78 @@
+/*******************************************************************************
+ * Copyright (c) 2020, 2022 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Eurotech - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.kapua.commons.util.qr;
+
+import com.google.zxing.BarcodeFormat;
+import com.google.zxing.WriterException;
+import com.google.zxing.client.j2se.MatrixToImageWriter;
+import com.google.zxing.common.BitMatrix;
+import com.google.zxing.qrcode.QRCodeWriter;
+
+import javax.imageio.ImageIO;
+import java.awt.Color;
+import java.awt.Graphics;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.net.URI;
+import java.util.Base64;
+
+public class QRCodeBuilderImpl implements QRCodeBuilder {
+    private static final int QR_CODE_SIZE = 134;  // TODO: make this configurable?
+    private static final String IMAGE_FORMAT = "png";
+
+    @Override
+    public String generateQRCode(URI uri) {
+        try {
+            // url to qr_barcode encoding
+            BitMatrix bitMatrix = null;
+            bitMatrix = new QRCodeWriter().encode(uri.toString(), BarcodeFormat.QR_CODE, QR_CODE_SIZE, QR_CODE_SIZE);
+            BufferedImage image = buildImage(bitMatrix);
+            return imgToBase64(image);
+        } catch (WriterException e) {
+            throw new RuntimeException(e);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Converts a {@link BufferedImage} to base64 string format
+     *
+     * @param img the {@link BufferedImage} to convert
+     * @return the base64 string representation of the input image
+     * @since 1.3.0
+     */
+    private static String imgToBase64(BufferedImage img) throws IOException {
+        final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+        ImageIO.write(img, IMAGE_FORMAT, outputStream);
+        return Base64.getEncoder().encodeToString(outputStream.toByteArray());
+    }
+
+    /**
+     * Converts a {@link BitMatrix} to a {@link BufferedImage}
+     *
+     * @param bitMatrix the {@link BitMatrix} to be converted into ad image
+     * @return the {@link BufferedImage} obtained from the conversion
+     * @since 1.3.0
+     */
+    private static BufferedImage buildImage(BitMatrix bitMatrix) {
+        BufferedImage qrCodeImage = MatrixToImageWriter.toBufferedImage(bitMatrix);
+        BufferedImage resultImage = new BufferedImage(QR_CODE_SIZE, QR_CODE_SIZE, BufferedImage.TYPE_INT_RGB);
+
+        Graphics g = resultImage.getGraphics();
+        g.drawImage(qrCodeImage, 0, 0, new Color(232, 232, 232, 255), null);
+
+        return resultImage;
+    }
+}
