+++ service/security/shiro/src/main/java/org/eclipse/kapua/service/authentication/shiro/realm/UserPassCredentialsMatcher.java
@@ -100,6 +100,4 @@ public class UserPassCredentialsMatcher implements CredentialsMatcher {
 
         return false;
     }
-
-
 }
